+++ include/accounting.h
@@ -1,0 +1,11 @@
+
+#ifndef ACCOUNTING_H
+#define ACCOUNTING_H
+
+int isIncoming(int opc);
+int isExpense(int opc);
+int isSeeListMovements(int opc);
+
+int isExit(int opc);
+
+#endif
